+++ src/plugins/azoth/plugins/sarin/filetransferin.h
@@ -68,7 +68,7 @@ namespace Sarin
 		void Accept (const QString&) override;
 		void Abort () override;
 	private slots:
-		void handleData (qint32, qint8, const QByteArray&);
+		void handleData (qint32, quint32, const QByteArray&);
 		void handleFileControl (qint32, qint32, int);
 	};
 }
